backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pydantic>=2.7.1
reportlab>=4.1.0
pillow>=10.4.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
python-dotenv>=1.0.1
httpx>=0.27.0
fpdf2==2.7.9
